Add the cart modal
Added a cart modal the user can interact with on all pages.
Access to said modal is gained via clicking the cart button in the navbar.
The cart will show all items added to it including additional information.
Note, the ability to add items to the cart was added in this commit.
Furthermore, this commit doesn't include any relevant tests.
Tests for the cart functionality will be added in a future commit.